feat: rework operation row to table columns with inline description
