+++ src/renderer/features/multisig-operations/components/OperationDescription.tsx
@@ -1,25 +1,17 @@
-import { useUnit } from 'effector-react';
-import { type ReactNode, useState } from 'react';
 import { Trans } from 'react-i18next';
-import { toast } from 'sonner';
 
 import { type Chain } from '@/shared/core';
 import { useI18n } from '@/shared/i18n';
 import { resolveDescriptionAreaState } from '@/shared/lib/operation-description/resolveDescriptionAreaState';
 import { Button, DetailRow, FootnoteText } from '@/shared/ui';
-import { Modal, TextArea } from '@/shared/ui-kit';
-import {
-  PERMISSIONS,
-  descriptionSaveErrorMessage,
-  operationDescriptionsResource,
-  operationsService,
-  useOperationDescription,
-} from '@/domains/backend';
+import { Modal } from '@/shared/ui-kit';
+import { useOperationDescription } from '@/domains/backend';
 import { type MultisigOperation } from '@/domains/network';
-import { contactModel } from '@/entities/contact';
-import { authModel, backendConfigurationModel, connectionHistoryModel } from '@/aggregates/backend';
-import { ReconnectAddressBookButton, backendContactsModel } from '@/features/contacts';
+import { ReconnectAddressBookButton } from '@/features/contacts';
 import { NamedAccount } from '@/widgets/NameResolver';
+import { useDescriptionEditing } from '../lib/use-description-editing';
+
+import { DescriptionEditorModal } from './DescriptionEditorModal';
 
 type Props = {
   operation: MultisigOperation;
@@ -27,104 +19,12 @@ type Props = {
 };
 
 const DESCRIPTION_PREVIEW_LENGTH = 40;
-type DescriptionModalMode = 'add' | 'edit';
 
 export const OperationDescription = ({ operation, chain }: Props) => {
   const { t } = useI18n();
   const description = useOperationDescription(operation.id);
-  const { authState, baseUrl, contacts, hasEverConnected, isHealthy } = useUnit({
-    authState: authModel.$authState,
-    baseUrl: backendConfigurationModel.$backendUrl,
-    contacts: contactModel.$backendContacts,
-    hasEverConnected: connectionHistoryModel.$hasEverConnected,
-    isHealthy: backendContactsModel.$isHealthy,
-  });
-
-  const [modalMode, setModalMode] = useState<DescriptionModalMode | null>(null);
-  const [draft, setDraft] = useState('');
-  const [isSaving, setIsSaving] = useState(false);
-
-  const hasWritePermission = authState?.permissions.includes(PERMISSIONS.OPERATION_WRITE) ?? false;
-  const isInAddressBook = contacts.some(contact => contact.accountId === operation.multisigAccountId);
-  const descriptionToSave = draft.trim();
-  const canSave = Boolean(baseUrl) && descriptionToSave.length > 0 && !isSaving;
-  const canEditDescription = Boolean(baseUrl) && isHealthy && hasWritePermission && isInAddressBook;
-
-  const handleCancel = () => {
-    setDraft('');
-    setModalMode(null);
-  };
-
-  const handleModalToggle = (open: boolean) => {
-    if (open) {
-      const nextMode = description ? 'edit' : 'add';
-      setDraft(description ?? '');
-      setModalMode(nextMode);
-    } else {
-      handleCancel();
-    }
-  };
-
-  const handleSave = async () => {
-    if (!baseUrl || !modalMode || descriptionToSave.length === 0) return;
-
-    setIsSaving(true);
-
-    try {
-      if (modalMode === 'edit') {
-        await operationsService.updateDescription(baseUrl, operation.id, descriptionToSave);
-      } else {
-        await operationsService.createDescription(baseUrl, {
-          multisigAccountId: operation.multisigAccountId,
-          chainId: operation.chainId,
-          callHash: operation.callHash,
-          blockNumber: operation.blockCreated,
-          extrinsicIndex: operation.indexCreated,
-          description: descriptionToSave,
-        });
-      }
-
-      operationDescriptionsResource.descriptionCreated({
-        id: operation.id,
-        description: descriptionToSave,
-      });
-      handleCancel();
-    } catch (error) {
-      const errorDescription = descriptionSaveErrorMessage(error, t);
-      toast.error(t('operation.descriptionSaveError'), { description: errorDescription });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const renderDescriptionEditorModal = (trigger: ReactNode) => (
-    <Modal size="mdlg" height="fit" isOpen={modalMode !== null} onToggle={handleModalToggle}>
-      <Modal.Trigger>{trigger}</Modal.Trigger>
-      <Modal.Title close>
-        {modalMode === 'edit' ? t('operation.editDescriptionTitle') : t('operation.addDescriptionButton')}
-      </Modal.Title>
-      <Modal.Content>
-        <div className="px-5 py-2">
-          <TextArea
-            value={draft}
-            placeholder={t('operation.descriptionPlaceholder')}
-            rows={8}
-            maxLength={500}
-            autoFocus
-            onChange={setDraft}
-          />
-        </div>
-      </Modal.Content>
-      <Modal.Footer>
-        <Button size="sm" variant="text" onClick={handleCancel}>
-          {t('operation.cancelDescriptionButton')}
-        </Button>
-        <Button size="sm" disabled={!canSave} isLoading={isSaving} onClick={handleSave}>
-          {t('operation.saveDescriptionButton')}
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  );
+  const { canEdit, hasWritePermission, isInAddressBook, isHealthy, hasEverConnected } =
+    useDescriptionEditing(operation);
 
   if (description) {
     const isLongDescription = description.length > DESCRIPTION_PREVIEW_LENGTH;
@@ -153,12 +53,16 @@ export const OperationDescription = ({ operation, chain }: Props) => {
               </Modal.Content>
             </Modal>
           )}
-          {canEditDescription &&
-            renderDescriptionEditorModal(
-              <Button size="sm" variant="text" className="shrink-0 p-0">
-                {t('operation.editDescriptionButton')}
-              </Button>,
-            )}
+          {canEdit && (
+            <DescriptionEditorModal
+              operation={operation}
+              trigger={
+                <Button size="sm" variant="text" className="shrink-0 p-0">
+                  {t('operation.editDescriptionButton')}
+                </Button>
+              }
+            />
+          )}
         </div>
       </DetailRow>
     );
@@ -209,11 +113,14 @@ export const OperationDescription = ({ operation, chain }: Props) => {
 
   return (
     <DetailRow label={t('operation.descriptionLabel')}>
-      {renderDescriptionEditorModal(
-        <Button size="sm" variant="text" className="p-0">
-          {t('operation.addDescriptionButton')}
-        </Button>,
-      )}
+      <DescriptionEditorModal
+        operation={operation}
+        trigger={
+          <Button size="sm" variant="text" className="p-0">
+            {t('operation.addDescriptionButton')}
+          </Button>
+        }
+      />
     </DetailRow>
   );
 };
